Refactor API and fix config
