+++ packages/bff/src/entities/Profile.ts
@@ -3,16 +3,10 @@ import { Column, CreateDateColumn, Entity, PrimaryColumn, UpdateDateColumn } fro
 @Entity()
 export class Profile {
   @PrimaryColumn()
-  id: string; // sid
+  sub: string;
 
   @Column({ length: 255, nullable: true })
   language: string;
-
-  @Column({ length: 255, nullable: true })
-  reportee: string;
-
-  @Column('simple-json', { nullable: true })
-  settings: any;
 
   @CreateDateColumn()
   createdAt: Date;
